pyproject.toml
[project]
name = "homelab"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-shellrunner>=0.3.5",
    "requests>=2.31.0",
]

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.3",
]

[tool.rye.scripts]
"lint:pyright" = "pyright ."
"lint:ruff" = "ruff check ."
lint = { chain = ["lint:pyright", "lint:ruff"] }
format = "ruff format ."

[tool.pyright]
typeCheckingMode = "strict"
extraPaths = ["lib"]

[tool.ruff]
line-length = 120
output-format = "grouped"
respect-gitignore = false

[tool.ruff.lint]
select = ["ALL"]
ignore = [
"D100",
"D101",
"D103",
"D203",
"D213",
"ANN101",
"T201",
"INP001",
# ruff format conflicts
"COM812",
"ISC001"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
